Changes made: To reorganize the work flow, we made the following changes - (1) llama_index_rag.py: added functions load_doc, load_docs, docs_to_nodes, persist_to_dir, (2) rag_agents.py: relocate internal function _prepare_args_from_config to llama_index_rag.py, as a method of class LlmaIndexRAG(RAGBase)
